requirements.txt
torch>=1.12.0
opencv-python
flask
numpy
pandas
imutils
pyyaml
pytest
